replay: change tuning progress type
As per request by tuning tools, change
the progress number to integer
64 bits and apply code style check.

+++ framework/tcpclient/tcp_client.h
@@ -137,7 +137,7 @@ class TcpClient
     }
 
     // Send the trace file pointer position over TCP
-    void TcpSendFilePos(double file_len, double bytes_sent, char* file_name);
+    void TcpSendFilePos(int64_t file_len, int64_t bytes_sent, char* file_name);
 
     // Send loaded driver info over TCP
     void TcpSendDriverLoadInfo();
